worker/requirements.txt
playwright>=1.40.0
edge-tts>=6.1.0
feedparser>=6.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
python-telegram-bot>=20.0
supabase>=2.0.0
google-genai>=1.63.0
youtube-transcript-api>=1.2.4
groq>=1.0.0
yt-dlp>=2026.2.0
